Restructure tests. Tests for factories. Replace usage of constructor of SudokuBoard by the factory methods.
