+++ src/solver/AbstractBoardFactory.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -37,7 +38,7 @@ public abstract class AbstractBoardFactory implements BoardFactory {
 
     protected abstract Collection<Collection<String>> makeUnitCollection();
     
-    protected Collection<String> crossArrays(char[] array1, char[] array2){
+    protected static Collection<String> crossArrays(char[] array1, char[] array2){
         Collection<String> crossedCollection = new ArrayList<String>(array1.length*array2.length);
         for(char elem1 : array1) {
             for(char elem2 : array2) {
@@ -47,7 +48,7 @@ public abstract class AbstractBoardFactory implements BoardFactory {
         return crossedCollection;
     }
 
-    protected Map<String, Collection<Collection<String>>> makeSquareUnitRelations(Collection<String> squares, Collection<Collection<String>> unitCollection) {
+    protected static Map<String, Collection<Collection<String>>> makeSquareUnitRelations(Collection<String> squares, Collection<Collection<String>> unitCollection) {
         Map<String, Collection<Collection<String>>> squareUnitRelations = new HashMap<String, Collection<Collection<String>>>();
         for(String square : squares) {
             Collection<Collection<String>> unitsOfSquare = new ArrayList<Collection<String>>();
@@ -61,7 +62,7 @@ public abstract class AbstractBoardFactory implements BoardFactory {
         return squareUnitRelations;
     }
     
-    protected Map<String, Collection<String>> makeSquarePeers(Collection<String> squares, Map<String, Collection<Collection<String>>> units) {
+    protected static Map<String, Collection<String>> makeSquarePeers(Collection<String> squares, Map<String, Collection<Collection<String>>> units) {
         Map<String, Collection<String>> peersPerSquare = new HashMap<String, Collection<String>>();
         
         for(String square : squares) {
@@ -80,7 +81,10 @@ public abstract class AbstractBoardFactory implements BoardFactory {
 
     protected Map<String, Collection<Character>> makeCurrentBoardFromInitialBoard(Map<String, Character> initialBoard) {
         Map<String, Collection<Character>> currentBoard = new HashMap<String, Collection<Character>>();
-        initialBoard.forEach((key,value) -> currentBoard.put(key, makeAllowedValues().toString().chars().mapToObj(e->Character.valueOf((char) e)).collect(Collectors.toList())));
+        List<Character> allAllowedValues = String.valueOf(makeAllowedValues()).chars().mapToObj(e->Character.valueOf((char) e)).collect(Collectors.toList());
+        initialBoard.forEach((key,value) -> {
+            currentBoard.put(key, allAllowedValues.stream().collect(Collectors.toList()));
+        });
         return currentBoard;
     }
